Relative path lang_def.xml + standalone mode
Changed the references to lang_def.xml in the source code
and added a static main method, so we can run the tool
in standalone mode.

Launching the editor now works in standalone mode,
but we still need to customize the controller to make it work
when packaged as a processing tool.

Added those changes:

- local reference to lang_def.xml inside BlocksTool
- local reference to lang_def.dtd in inside lang_def.xml
- static main method taking lang_def.xml location as command line
  argument